long double model

+++ src/model_s.cpp
@@ -5,16 +5,15 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-
-ModelS::ModelS(double s){
+ModelS::ModelS(long double s){
     strike = s;
 }
 
 // safe calculation
-double ModelS::safePower(double base, int power){
+long double ModelS::safePower(long double base, int power){
     feclearexcept(FE_OVERFLOW);
     feclearexcept(FE_UNDERFLOW); // overflow and underflow detection
-    double ret = pow(base, power);
+    long double ret = powl(base, (long double)power);
     if((bool)fetestexcept(FE_OVERFLOW)
               || (bool)fetestexcept(FE_UNDERFLOW)){
         throw std::runtime_error("power raise exception!\n");
@@ -48,7 +47,7 @@ state ModelS::indexToState(int idx){
     }
     result.period = period;
     // calculate state stock price
-    double price = initialPrice * safePower(up, period);
+    long double price = initialPrice * safePower(up, period);
     int diff = idx - start;
     for (int c = 0; c < diff; c++){
         price = price / up * down;
@@ -57,18 +56,18 @@ state ModelS::indexToState(int idx){
     return result;
 }
 
-double ModelS::worker(int idx){
+long double ModelS::worker(int idx){
     //printf("here is worker working on idx %d \n", idx);
     state now = indexToState(idx);
     if(now.period == MAXPERIOD){
         // leaf of tree model calculate option value
-        return std::max(now.stockPrice - strike, 0.0);
+        return std::max(now.stockPrice - strike, 0.0L);
     }
     else
     {
         int head = nextUp(idx, now);
         int tail = head + 1;
-        printf("time period: %d, up %f  down %f \n", now.period, buffer[head], buffer[tail]);
+        printf("time period: %d, up %Lf  down %Lf \n", now.period, buffer[head], buffer[tail]);
         return 1.0/(1+rate)*(p*buffer[head] + q*buffer[tail]);
     }
 }
@@ -79,14 +78,14 @@ int ModelS::nextUp(int idx, state s){
     return begin + diff;
 }
 
-double ModelS::calculate(double u, double d, double r, double s0){
+long double ModelS::calculate(long double u,long double d,long double r,long double s0){
     up = u;
     down = d;
     rate = r;
     initialPrice = s0;
     p = (1 + rate - down) /(up - down);
     q = (up - rate - 1) /(up - down);
-    printf(" p: %f  q: %f  \n", p, q);
+    printf(" p: %Lf  q: %Lf  \n", p, q);
     for(int i = (MAXPERIOD+1)*(MAXPERIOD+2)/2 - 1; i >= 0; i--){
         buffer[i] = worker(i);
     //    printf("Finish Calculating %d \n", i);
